requirements.txt
pyserial
ephem
Pillow
scapy
numpy
sklearn